staging/eidolon/requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.25.0
datasets>=2.14.0

# Computer vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
lxml>=4.9.0
